将类和函数分别放到两个头文件中，减少主文件的代码量
